Clean up the logging, improve naming, remove redundant dependencies.
