+++ client-angular/src/app/app.component.ts
@@ -39,7 +39,6 @@ export class AppComponent implements OnInit {
 
   ngOnInit(): void {
     this.navService.config$.subscribe(config => {
-      console.log(config);
       this.showNav = config.showNavigation;
       this.toolbarLabel = config.toolbarLabel;
     });
