--- client-angular/src/app/layout/navigation/navigation.component.ts
@@ -1,12 +1,0 @@
-import {Component} from '@angular/core';
-
-/**
- * A component responsible for displaying navigation bar on the left of the page.
- */
-@Component({
-  selector: 'app-navigation',
-  templateUrl: './navigation.component.html',
-  styleUrls: ['./navigation.component.css']
-})
-export class NavigationComponent {
-}
