+++ client-angular/src/app/task-creation-wizard/step-2-label-assignment/label-assignment.component.ts
@@ -67,11 +67,8 @@ export class LabelAssignmentComponent extends WizardStep {
    */
   private readonly loadAvailableLabels: Promise<LabelView[]>;
 
-  constructor(router: Router,
-              wizard: TaskCreationWizard,
-              labelService: LabelService,
-              layoutService: LayoutService) {
-    super(router, wizard, layoutService);
+  constructor(router: Router, wizard: TaskCreationWizard, labelService: LabelService) {
+    super(router, wizard);
     this.loadAvailableLabels = labelService.fetchAllLabels();
   }
